ci: improve merge commit content
This commit introduces a commit message template in the mergify.yml
configuration file. The template includes the pull request title,
number, body, and a list of users who approved the review.

One important feature is to add the name of approvers to the PR, so the
merge commit will include fields like:

```
Approved-by: leseb
```

Signed-off-by: Sébastien Han <seb@redhat.com>