Add auto max_workers capability
This capability takes advantage of the feature from the eval library which tunes max_workers according to the hardware configuration.  In order to accomplish this, effective gpus and max_workers="auto" needs to be passed to eval.  To accomplish this, gpus and effective gpus needed to be calculate earlier in the process.

This change also updates training to use get_gpus and max_workers=auto.  This means training now uses the gpu settings of evaluate when calling evaluate before defaulting to all gpus if not specified.

Signed-off-by: Dan McPherson <dmcphers@redhat.com>