map total vram -> profiles
each profile should have a "max vram" amount. If our vram is < that maximum during a loop, then that is the profile we want.

We basically default the user to the most minimal profile possible: if they have 8GB they default to our 16 GB profile, if they have 96, they default to the 160 2xA100 profile, etc.

Signed-off-by: Charlie Doern <cdoern@redhat.com>