fix: mypy linting on MacOS
Mypy was failing to lint classes using the mlx dependency because mlx is
conditionally imported only on macOS. Since mypy performs static
analysis, it requires access to all imported modules to analyze their
types. Running mypy in an environment without mlx installed (e.g., on
    Windows or Linux) resulted in no linting errors while failing on
MacOS.

To avoid failing on MacOS we added a few files to the exclude list, so
these will be skipped by mypy when running.

Signed-off-by: Sébastien Han <seb@redhat.com>