feat: use default and description from Pydantic model
Our Pydantic model, referred to internally as Config, serves as the
source of truth for default values and descriptions. When a Click CLI
option is also present in the Config, it should inherit its default
value and description from the corresponding Config field. When applying
the cls=clickext.ConfigOption configuration to a Click option, fields
such as “default” and “help” should not be specified, as these will
automatically be derived from the Config object.

So this is how you should write an option that exists in the Config
(ultimately in config.yaml) AND has a CLI option:

```
my_config_option: StrictStr = Field(
    default="my_default_value",
    description="Great description of what the option does",
)

@click.option(
    "--my-config-option",
    type=click.STRING,
    cls=clickext.ConfigOption,
)
```

That's all. Later we could try to do the same for the type, so we don't
need `type=click.STRING`. It's a bit more tedious and is a larger
change.

Signed-off-by: Sébastien Han <seb@redhat.com>