feat: train profile auto selection
add auto selection for NVIDIA GPU train profiles. This uses torch under the hood
to read the amount of GPUs, tally the total vram, the GPU type, and the GPU amount. Use this information to choose the train profile which best matches the amount of GPUs/the GPU type. If none of these match, then match based off of vRAM. for good measure, set nproc_per_node at the end if basing of off vRAM. in case we found a profile which matches our vRAM amount but not our GPU amount

Signed-off-by: Charlie Doern <cdoern@redhat.com>