feat: move model list logic to utils.py
We currently handle all operations within the implementations of the CLI commands
themselves. This commit begins to refactor this by moving the implementation of
ilab model list to have the raw machinery exist in `instructlab.utils`.

Due to how the current dependencies are arranged, we also had to move the
`is_model_gguf` and `is_model_safetensors` functions to exist in the
utils.py file as well, since moving the list logic to utils.py in the current
state causes us to encounter circular dependency issues.

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>