feat: make flash-attention toggle a first-class option
Previously, the `disable_flash_attn` training flag was set as an additional argument that was hidden
away. However; we run into problems when running E2E testing in training where it cannot run on a T4
system (Volta GPU). This commit resolves that issue by promoting `disable_flash_attn` as a
first-class argument that can be passed as a CLI argument

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>