cloud-instance.sh: Add ssh config management
When using stop/start to turn off and on dev instances, the public IP
address will change. I keep an entry in my ssh config to make it easy
to connect to these instances (both with ssh manually, as well as with
vscode over ssh). I found myself using "ec2 details" to get the new IP
and then updating my config. After a handful of times of this, I
decided to automate it.

Once an instance is running, you can run `ec2 update-ssh-config` and
it will either add the instance to the ssh config, or update it with
the latest address (via the hostname derived from the public IP, since
there was already code in the script to grab that).

The change should work on both ec2 and ibm cloud, though I've only
tested this with ec2.

Signed-off-by: Russell Bryant <rbryant@redhat.com>