vllm: Add support for specifying vllm path
This change adds a new configuration option for vllm: a path to the
vllm exxecutable. This allows you to point to an instance of `vllm`
installed in a different Python virtual environment than `ilab`.

This also changes the `vllm` backend to run `vllm serve` instead of an
older style of specifying the openai api server entrypoint manually.

As an example of how this is useful, I wanted to try the following
vllm PR to run mixtral on a host with less GPU memory than is
typically required. Running this version has Python dependencies not
compatible with current `ilab`.

    https://github.com/vllm-project/vllm/pull/8751

To install vllm into its own venv:

    python3.11 -m venv venv-vllm
    . venv-vllm/bin/activate
    git clone https://github.com/vllm-project/vllm
    cd vllm
    git fetch origin pull/8751/head
    git checkout FETCH_HEAD
    export CUDA_HOME=/usr/local/cuda
    export PATH=${CUDA_HOME}/bin:${PATH}
    pip install -e .

Back in my instructlab venv ...

    . venv-ilab/bin/activate
    ilab model download --repository mistralai/Mixtral-8x7B-Instruct-v0.1
    ilab config edit

Edit vllm config. Note the value of `--tensor-parallel-size` is for
the number of GPUs I have available.

    serve:
      backend: vllm
      model_path: /home/ec2-user/.cache/instructlab/models/mistralai/Mixtral-8x7B-Instruct-v0.1
      vllm:
        vllm_args: ['--tensor-parallel-size', '4', '-q', 'fp6_weights']
        vllm_path: "/home/ec2-user/venv/bin/vllm"

This now works as mixtral is quantized at load-time, cutting the gpu
memory requirement in half.

Signed-off-by: Russell Bryant <rbryant@redhat.com>