implements higher-granularity bookkeeping during training
enables restart in the event of a system crash, GPU hang, etc.

Adds a TrainingJournal object that represents a JournalModel pydantic
class. During phased training, an instance of the JournalModel is
updated and written to disk as progress is made.

In the event of an untimely exit from the training process, a
'journalfile.yaml' by convention, or a 'yaml' file by another name, can
be loaded into a "JournalModel" object and training can resume.

Signed-off-by: James Kunstle <jkunstle@redhat.com>