fix: two phased training fixes
Phased training is broken, and this commit has
the following fixes for it:

1. Phased training does not require a valid
data-dir. It has phased training flags and config
for the data-dir in phase 1 and phase 2.

2. The eval section from the config was trying
to be read from ctx.obj, not ctx.obj.config.
This lead to errors saying the ctx.obj does
not have a "evaluate" member.

Signed-off-by: Ali Maredia <amaredia@redhat.com>