e2e: Restore using pull_request_target target for e2e workflow
In #2279, we split the e2e workflow into a starter pull_request+push
target workflow which passed metadata to a
workflow_run target workflow.
This caused some issues in visibility of the e2e
jobs to the PR check-runs list and to mergify for the auto-merge config.

This PR restores the use of pull_request_target while retaining the
improvements for using the status-checks workflow to defer execution
until jobs from other workflows have successfully completed.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>