Add step to fix Python version management tool issue
Some Python version management tools may not by default build libraries
implemented in C, and CPython when installing a Python version. This can
result in missing libraries in the Python install. This can be fixed by
requesting the C, and CPython libraries to be explicitly built
using `--enable-framework` flag.

Fixes #2347

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>