fix: stabilize checkpoint selection logic
The current implementation of checkpoint selection has a bug which is a result of the `sorted` function
not being able to correctly compare strings which all have an equal prefix and a numeric suffix.
This commit fixes that bug by explicitly telling `sorted` how to perform the comparison between items.

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>