e2e: Use fork of wait-for-status-checks action with needed fix
We need the fix to support pull_request_target target workflows
properly. A bug in the wait-for-status-checks action looks at the
wrong commit SHA (the target branch, e.g. main, instead of
pull/nnn/head). The bug causes the ready check to look at the jobs
from the last commit on the target branch not the last commit on
the PR (pull/nnn/head).

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>