Update Python 3.12 and PyTorch 2.4.1
Now that PyTorch 2.4 supports Python 3.12, This change attempts updating
the default Python version to 3.12. It also brings the update to PyTorch
2.4.1 and all its dependencies. One of these dependencies is AMD ROCm
6.1. So, it ends up being a massive change.

vLLM also needs to be update to v0.6.x.
Intel Gaudi (HPU) remains on Intel fork of PyTorch 2.3.1.

It also requires the other repositories to update their dependencies
accordingly, so it requires some coordination.

Fixes #1463

Signed-off-by: Fabien Dupont <fdupont@redhat.com>