fix: switch vram approximation
vram approximation should be >= not <. Using less than can lead to situations where we overmatch to a profile that can cause OOM situations.
Making sure our profile is "overfit" for the vRAM makes more sense. However, these has to be a threshhold here so we don't overfit in the other direction

Signed-off-by: Charlie Doern <cdoern@redhat.com>