chore: wrap training library with try/catch handles
Previously, the instructlab/training run_training function would handle errors
by silently failing and reporting a success. As of training PR #217, we have fixed
this behaior by now raising a RuntimeError when we encounter a problem inside of
the training loop. This commit handles this error inside of the accelerated and
phased training pipelines.

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>