add system profile auto detection
detect system profiles based off of exact matches only. If a user is on an Apple M3 Pro system, we will match that to the name of the profile we write to disk.
The same goes for all Nvidia system profiles currently supported.

We are not doing best approximation here, at least not yet. The approximation proved to add confusion with training profiles. We also cover many more use cases with system profiles than we did with training profiles so odds are you should have some sort of direct match!

Signed-off-by: Charlie Doern <cdoern@redhat.com>