Adding library syncing for instructlab repos
Supports syncing for the sdg, training, and eval repos.  Repos are cloned onto the remote instance and installed with pip -e.  When synced, the delta is sent to the remote instance with git and extracted so it will take effect with the pip -e installed library.

This enables a workflow where instances are started and synced to from your local environment without having to setup an environment on the remote instance.

Signed-off-by: Dan McPherson <dmcphers@redhat.com>