Fix Intel Auto-Detection and Profile Selection Menu
Intel returns Intel(r) for their CPUs, so we need to do a more general check than strict matching `intel`.

Also, add a [0] option for profiles that just sets the pydantic defaults. It is unclear who would get into this scenario with the auto-detection working
but if a user is running on such a system, the pydantic defaults might just be what they're looking for

Signed-off-by: Charlie Doern <cdoern@redhat.com>