fix: add upsampling as an option
In SDG today, when user-generated knowledge data is mixed in with skills it is done
without consideration for the difference in size between knowledge and skills.
As a result, when the skills dataset greatly outnumbers knowledge by several orders of magnitude,
we see model degradation in terms of knowledge learned. This commit resolves that by
allowing the CLI to accept an `upsample_amount` which controls how much of the knowledge
we upsample into the CLI. I.e, how many samples of the knowledge do we need in the skills dataset
for it to not forget. This is implemented in SDG by simply replicating those samples up to a given
number N. This commit simply exposes this setting in the CLI.

For example, if you had a knowledge dataset which you generated and was 20 samples in length,
specifying an upsample_amount of 100 would cause the data to be duplicated until it is 100 samples in length.
That is - the resultant dataset would have (ideally) 5 copies of the original dataset.

Signed-off-by: Oleg S <97077423+RobotSail@users.noreply.github.com>