switch auto-detection to use metadata
profile selection should read the metadata from the profile rather than matching the profile based on the file name
this switch also involves checking if the name of the GPU on the system matches any of the SKU info in the metadata.
so if you are on a Nvidia H100 PCIe system as detected by torch we will check if your system matches Nvidia H100 NVL, Nvidia H100 PCIe, or Nvidia H100.
This ensures we can match multiple SKUs of the same GPU to a single config profile

Signed-off-by: Charlie Doern <cdoern@redhat.com>