Require `sdg>=0.6.2` + pin `pydantic<=2.9.2`
`docling-parse` v3 rearranged a lot of logic and deprecated certain syntaxes, resulting in InstructLab failing to build/run. This issue was resolved in SDG release v0.6.2. Therefore, we want to require `sdg>=0.6.2`.

Also, `pydantic` v2.10.0 introduces breaking changes that affect `mypy`, which is used for our linting steps. Therefore, I'm pinning to `pydantic` v2.9.2 with `mypy` to v1.13 (or less) to avoid this issue.

Finally, add `requirements-dev.txt` as a list of default dependencies for the tests to ensure all requiured packages are being used for testing.

Signed-off-by: Courtney Pacheco <6019922+courtneypacheco@users.noreply.github.com>