Pin `docling-parse` to latest v2 + pin `pydantic<=2.9.2`
In v3, `docling-parse` rearranged a lot of logic and deprecated certain syntaxes. For now, we will pin to v2 while we investigate updating our syntax to use the new v3 version's syntax.

Also, `pydantic` v2.10.0 introduces breaking changes that affect `mypy`, which is used for our linting steps. Therefore, I'm pinning to `pydantic` v2.9.2 with `mypy` to v1.13 (or less) to avoid this issue.

Finally, add `requirements-dev.txt` as a list of default dependencies for the tests because some tests need to reference the newly-hardcoded `docling-parse` version. The default dependenices can be overridden.

Signed-off-by: Courtney Pacheco <6019922+courtneypacheco@users.noreply.github.com>