version 0.3.5 of llama-cpp-python has a known issue version 0.3.2 has granite 3.0 support and does not have this issue. Bump to this version
this required some additions to how we handle chat exceptions. As of these newer 0.3.z llama-cpp-python versions,
a bad request causes the server to die. This requires us to know the max_ctx_size of the server before passing a completions request so that
we can maintain the behavior of trimming messages until we can respond to one that fits.

in order to do this, the config now contains a `current_max_ctx_size` field that we will update when spinning up a server.
in the case that a user implicitly starts a llama-cpp-python server when calling `ilab model chat`, we set the max_tokens to the
current `max_ctx_size` in the serve config.

Signed-off-by: Charlie Doern <cdoern@redhat.com>
(cherry picked from commit b29efddadaa731a15c4bd2f2494a0b6731db6c29)
Signed-off-by: Charlie Doern <cdoern@redhat.com>