chore!: Update vLLM to 0.6.6.post1
This PR is a follow-up to #2865 that relaxed the PyTorch version range.
Even with that range extension, we realized that PyTorch 2.4 is still
used when installing `instructlab[vllm-cuda]`, because vLLM 0.6.2 has a
requirement on PyTorch 2.4.

This new PR updates the version of vLLM to 0.6.6.post1, which is the
latest available in the Open Data Hub fork of vLLM. The vLLM changelog
doesn't highlight much risk in this version bump.

Resolves #2702

Signed-off-by: Fabien Dupont <fdupont@redhat.com>