feat: proper process listing and status
`ilab process list` was mis-representing process start times and did not have a status.

The output will not contain whether a process is running or finished. The runtime will also stop properly when the process is finished, in attached mode, in detached mode the Runtime will be stopped when the user lists the finished process

Signed-off-by: Charlie Doern <cdoern@redhat.com>