fix: process listing, status, and runtime
currently, we pop items off of the process registry while listing it causing an error. Remove all processes at the end of listing and mark them for removal properly
additionally, list_processes used to manage setting the status of a process. While this is easy, it doesn't quite make sense for a listing function to be where status is getting updated.

Instead, the target function should, when it errors out or exits, load and update the registry using the new `update_status`. The only other place we now manually manage status is inside of `stop_process`. Since we kill processes especially when running in attached mode, we cannot depend on the target to have time to update the registry.

Signed-off-by: Charlie Doern <cdoern@redhat.com>