feat: Remove hpu (Gaudi) support from simple training
Intel Gaudi support for simple training was very experimental and
limited. It was never properly tested and flagged as experimental from
the beginning. The design of the simple training code also did not allow
for multi-device training. Typical Gaudi servers have 4 or 8 devices.

The removal of Gaudi support makes it possible to remove dependency on
`optimum` and `optimum-habana` packages. The packages have upper version
limits for packages like `trl`.

Fixes: #2963
Signed-off-by: Christian Heimes <cheimes@redhat.com>