ci: Add DK-Bench into e2e-ci.sh
This PR does the following:
- Adds running of dk_bench into e2e-ci.sh
- Add test questions for dk-bench
- Adds the OPENAI_API_KEY to the x-large job

The addition of DK-Bench requires OPENAI_API_KEY
to be set before e2e-ci.sh is run.

If this key is not set, the program will exit with
a warning.

If a user of the script does not have an OPENAI_API_KEY
and wishes to run the script without running DK-Bench
they make set OPENAI_API_KEY='NO_API_KEY'.

Signed-off-by: Ali Maredia <amaredia@redhat.com>