tests: Fix failure on python without _uuid module
On these pythons, uuid module will fall back to a platform specific
method to extract a safe uuid. It usually means a call to an external
utility like `ip` or `ifconfig` (to derive a UUID from the machine MAC
address):

https://github.com/python/cpython/blob/379ab856f59423c570333403a7d5d72f3ea82d52/Lib/uuid.py#L641

Because of that, the Popen mock catches a spurious `ip` call, which
makes the later check fail because the observed number of calls is off
by 1. (The `uuid.uuid1()` that triggers it in the test suite comes from
process management layer.)

One example of python build with _uuid module is linux nixpkgs cpython.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>