Update NVIDIA CUDA to 12.8
NVIDIA Blackwell GPU is the newest NVIDIA accelerator and is hitting the
market in 2025. CUDA 12.8 adds the support for NVIDIA Blackwell, so this
change updates CUDA version to 12.8.

CUDA versions are backward compatible, so we don't expect any issue with
existing code.

The Blackwell support is added in NVIDIA drivers R570, so we also update
the drivers version to 570.86.15.

Signed-off-by: Fabien Dupont <fdupont@redhat.com>