Add new in-house GitHub action to detect secrets exposure
This GitHub action detects any secret accidentally exposed through a GH workflow that auto-triggers on PR builds. The action's logic relies on a "deny list" approach when deciding if a workflow is problematic or not.

Also, isort is disabled for the new GitHub action because it tries to add random comments to the Python scripts.

Signed-off-by: Courtney Pacheco <6019922+courtneypacheco@users.noreply.github.com>