Use prebuilt vLLM for NVIDIA CUDA
We have seen timeouts in the CI jobs because vLLM build takes too much
time. Building vLLM is only done because we use the Open Data Hub fork
of vLLM. However, the main difference between ODH fork and upstream vLLM
is the Dockerfile for UBI, which is not used in InstructLab.

This change modifies the requirements to use the prebuilt vLLM wheel
from PyPI, so that we don't build vLLM from source. This should speed up
the CI jobs in general.

Fixes #3111.

Signed-off-by: Fabien Dupont <fdupont@redhat.com>