ci: Remove spelling checks
I find them not useful and mostly distracting, esp. to new contributors.

Of course, this is not an endorsement to stop carrying about egregious
spelling issues, or where it's important (in user facing docs,
changelog, etc.) Even then, some of these could be handled in particular
moments in release schedule (when prepping a new release cut).

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>