infra: bail early on package install failures
When you specify multiple packages to DNF in RHEL 9, if some are
available and some are missing, DNF will print a warning and continue
on. As a result, the install-nvidia.sh script can proceed without
installing all dependencies, and when the user reboots, the nvidia
drivers may not function.

Reconfigure DNF to error on any missing packages. This helps us catch
problems sooner.

Combine the two dnf config-manager commands into one invocation at the
top of the script so the settings take effect for all subsequent DNF
operations.

Remove the "--best --nodocs" options, because they only affect dnf
install operations, not config-manager.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>