Add `return_dict_in_generate=True` to `model.generate()`
In certain situations, `linux_train.py` fails with: "UserWarning: `return_dict_in_generate` is NOT set to `True`, but `output_logits` is. When `return_dict_in_generate` is not `True`, `output_logits` is ignored."

To fix this failure, I have added `return_dict_in_generate=True` as an input to `model.generate()`.

Signed-off-by: Courtney Pacheco <6019922+courtneypacheco@users.noreply.github.com>