feat: Handle lists of pydantic models when dumping to yaml
Before the change, lists of models were not converted to dicts before
passing them to ruamel representor. It broke `config show` when a
`models` section was added to it because the field is s list of
model_info classes.

With this change, we should correctly handle tuples and lists. If we
ever need something else, like dicts, we may expand the serialization
code further.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>