use constraints-dev.txt in e2e tests
For instructlab, "pip install ." does not install vllm, but it does
install an uncapped torch (2.7.0 currently).

When we install vllm later, we compile a binary flash_attn wheel against
torch 2.7.0. vllm 0.8.4 requires torch==2.6.0, so we downgrade torch,
and then we use that with the incompatible flash_attn binary wheel.

To resolve this, use constraints-dev.txt in the first pip install
operation. This restricts torch to 2.6.0 immediately when we first
install instructlab, so that we will compile flash_attn against that
torch version.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>