Ignore unraisable exceptions from docling tests
Newer versions of Docling can raise an exception during custom garbage
collection code of TesseractOcrModel that nothing ever catches,
because exceptions during garbage collection are generally not caught
by anything. The exception is harmless to any of our InstructLab
use-cases, and only happens when the TesseractOcrModel was not able to
construct itself anyway, which we check for and handle to fallback to
EasyOCR or other implementations.

However, py3-unitcov by default fails the test suite if any of these
unraisable exceptions happen. This adjusts that to not fail the test
suite. Messages still get logged about the exception during regular
py3-unit testing, so we can see when this gets fixed in Docling, but
there's no need for it to actually fail our test suite.

Signed-off-by: Ben Browning <bbrownin@redhat.com>
(cherry picked from commit 65c1efb9ca538e90f841d6c83f7ef1750e3f4274)