ci: enable workflow_dispatch for all e2e jobs, part 1
This allows to test workflow changes to these jobs before merging them.

Note: this change doesn't include more risky changes like adding inputs
for pr number etc. This is intentional and will be done in a separate PR
when the Run workflow button is available in the UI and we can test
changes to workflows before merging.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>