ci[cuda]: build project from local checkout, not from PyPI
This makes it possible to build the CUDA container with local changes,
easing the developer workflow.

It also allows CI to actually build the container with the changes
proposed in a PR, so that we can detect when a change breaks the build
before it's merged rather than when it makes it into a release.

A new PIP_CONSTRAINT argument allows respecting constraints files. This
is disabled by default when using `make cuda` or building the
Containerfile directly, but we enable it in our CI job for more
stability.

Resolves #3339
Partially addresses #3299

Signed-off-by: Miguel Garcia <migarcia@redhat.com>