chore: don't min-cap pydantic specifically for mypy job
We should test with what is actually used in other CI jobs. If we have
an issue with type checker, we can fix it, ignore or - at worst - cap
the library in general until we solve the issue and uncap back.

Ignoring existing violations though we should probably revisit them.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>